Move to Type-system, implement review recommendations
